+++ src/js_realm.hpp
@@ -1045,33 +1045,43 @@ void RealmClass<T>::subscribe_to_objects(ContextType ctx, ObjectType this_object
     args.validate_count(3);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
-    auto class_name = Value::validated_to_string(ctx, args[0]);
-    auto query = Value::validated_to_string(ctx, args[1]);
+    std::string object_type = Value::validated_to_string(ctx, args[0]);
+    std::string query = Value::validated_to_string(ctx, args[1]);
     auto callback = Value::validated_to_function(ctx, args[2]);
+
+    auto &schema = realm->schema();
+    auto object_schema = schema.find(object_type);
+
+    if (object_schema == schema.end()) {
+        throw std::runtime_error("Object type '" + object_type + "' not found in schema.");
+    }
 
     Protected<ObjectType> protected_this(ctx, this_object);
     Protected<typename T::GlobalContext> protected_ctx(Context<T>::get_global_context(ctx));
+    Protected<FunctionType> protected_callback(ctx, callback);
     auto cb = [=](realm::Results results, std::exception_ptr err) {
+        HANDLESCOPE
+
         if (err) {
             try {
                 std::rethrow_exception(err);
             }
             catch (const std::exception& e) {
-                typename T::Value arguments[2];
-                arguments[0] = Value::from_null(protected_ctx);
-                arguments[1] = Value::from_string(protected_ctx, e.what());
-                Function<T>::callback(ctx, callback, protected_this, 2, arguments);
+                ValueType callback_arguments[2];
+                callback_arguments[0] = Value::from_string(protected_ctx, e.what());
+                callback_arguments[1] = Value::from_null(protected_ctx);
+                Function<T>::callback(ctx, protected_callback, protected_this, 2, callback_arguments);
             }
             return;
         }
 
-        typename T::Value arguments[2];
-        arguments[0] = ResultsClass<T>::create_instance(protected_ctx, results);
-        arguments[1] = Value::from_null(protected_ctx);
-        Function<T>::callback(protected_ctx, callback, protected_this, 2, arguments);
+        ValueType callback_arguments[2];
+        callback_arguments[0] = Value::from_null(protected_ctx);
+        callback_arguments[1] = ResultsClass<T>::create_instance(protected_ctx, results);
+        Function<T>::callback(protected_ctx, protected_callback, protected_this, 2, callback_arguments);
     };
 
-    partial_sync::register_query(realm, class_name, query, std::move(cb));
+    partial_sync::register_query(realm, object_type, query, std::move(cb));
 }
 #endif